+++ components/sidebar/SidebarElement.vue
@@ -1,6 +1,5 @@
 <script lang="ts" setup>
 import { NuxtIconName } from "#app";
-import { useAuth0 } from "@auth0/auth0-vue";
 import { RoutesNamedLocations } from "~/.nuxt/typed-router/__routes";
 import { version } from "~/package.json";
 
@@ -15,14 +14,6 @@ const links: {
 ];
 
 const { data: collections } = usePrivatePlaylists();
-const auth0 = useAuth0();
-const logout = async () => {
-  try {
-    await auth0.logout({ openUrl: false });
-  } catch (e) {
-    console.error(e);
-  }
-};
 </script>
 
 <template>
@@ -43,15 +34,6 @@ const logout = async () => {
           :key="`${link.link}_${i}`"
           v-bind="link"
         />
-        <a
-          @click="logout()"
-          href="#"
-          class="group flex gap-2 rounded-xl px-4 py-2"
-        >
-          <span class="transition-transform group-hover:translate-x-2">
-            Logout
-          </span>
-        </a>
       </SidebarGroup>
 
       <SidebarGroup title="Playlists">
